\name{colours_GUI}
\alias{colours_GUI}
\title{Colours for pRolocGUI}
\usage{
colours_GUI(transparency=70)
}
\arguments{
  \item{transparency}{Sets the transparency of plotted
  elements. Between 0 and 100 and of type \code{numeric}.
  Default is 70 as used in
  \code{\link{pRoloc::pRoloc-Tutorial}}.}
}
\description{
Sets the default colours as used in the package
\code{pRoloc}.
}
\examples{
\dontrun{colours_GUI(transparency=70)}
}
\author{
Thomas Naake <thomasnaake@gmx.de>
}
\seealso{
\code{\link[pRoloc]{setStockcol}}
}

