\name{digestFOI readSR descrFOI}
\alias{.descrFOI}
\alias{.digestFOI}
\alias{.readSR}
\alias{digestFOI readSR descrFOI}
\title{Helper functions for \code{reactivePoll} in \code{pRolocGUI}}
\usage{
.digestFOI()
}
\description{
\code{.digestFOI} is used as the \code{checkFunc} argument,
\code{.readSR} and \code{.descrFOI} as \code{valueFunc} in
\code{reactivePoll}
}

