% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotMat2D.R
\name{plotMat2D}
\alias{plotMat2D}
\title{Plot a matrix of markers}
\usage{
plotMat2D(object, fcol = "Markers", ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{fcol}{The name of the markers matrix. Default is
\code{"Markers"}.}

\item{...}{Additional parameters that can be used to choose the
dimentionality reduction method, as defined in
\code{\link{plot2D}}.}
}
\description{
Shiny App for visualising a matrix of markers
}
\examples{
library("pRoloc")
library("pRolocdata")
data(dunkley2006)
## adds matrix markers
dunkley2006 <- mrkVecToMat(dunkley2006)
plotMat2D(dunkley2006)
}
\author{
Lisa M Breckels
}

