\docType{package}
\name{pRolocGUI-package}
\alias{pRolocGUI-package}
\title{An interactive visualisation tool for organelle (spatial) proteomics data}
\description{
The package pRolocGUI comprises functions to interactively
visualise organelle (spatial) proteomics data on the basis
of pRoloc, pRolocdata and shiny.
}
\details{
\tabular{ll}{ Package: \tab pRolocGUI \cr Type: \tab
Package \cr Version: \tab 0.1.1 \cr Date: \tab 2014-04-28
\cr License: \tab GPL-2 \cr }
}
\examples{
\dontrun{pRolocGUI()}
}
\author{
{Thomas Naake and Laurent Gatto \cr Maintainer: Thomas
Naake <thomasnaake@gmx.de> }
}
\seealso{
\code{\link[pRoloc]{pRoloc-tutorial}}
\code{\link[pRolocdata]{pRolocdata}}
\code{\link[MSnbase]{FeaturesOfInterest-class}}
}

