% Generated by roxygen2 (4.0.0): do not edit by hand
\name{pRolocVis}
\alias{pRolocVis}
\title{pRolocVis}
\usage{
pRolocVis(object = NULL)
}
\arguments{
\item{object}{Pass a MSnSet to pRolocVis directly. Default \code{NULL} will
enable possibility to upload MSnSet in pRolovVis.}
}
\description{
A function to start a shiny session with one MSnSet data set.
Run \code{pRolocVis()} to start the shiny application and choose between
three example MSnSet originating from \code{pRolocdata} or upload your
own MSnSet. Choosing between the tabs allows to display PCA plots,
protein profiles, the underlying data and upload abilities for past
search results.
}
\examples{
\dontrun{pRolocVis(object = NULL)}
}
\author{
Thomas Naake <thomasnaake@gmx.de>
}

