% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pRolocVis2.R
\name{pRolocVis2}
\alias{pRolocVis2}
\title{Visualise your pRoloc data}
\usage{
pRolocVis2(object, fcol, foi, fig.height = "600px", fig.width = "100\%",
  legend.width = "100\%", legend.cex = 1, nchar = 15, all = FALSE, ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}.}

\item{fcol}{The name of the markers matrix (default is
\code{"Markers"}). Can be missing if \code{foi} is available.}

\item{foi}{A \code{\link{FeaturesOfInterest}} or a
\code{\link{FoICollection}}, that will be available for display.}

\item{fig.height}{Height of the figure. Default is \code{"600px"}.}

\item{fig.width}{Width of the figure. Default is \code{"600px"}.}

\item{legend.width}{Width of the legend. Default is \code{"100\%"}.}

\item{legend.cex}{Character expansion for the vignette
labels. Default is 1.}

\item{nchar}{Maximum number of characters if the markers class
names, before their names are truncated. Default is 10.}

\item{all}{If there are more than 10 clusters, only the first
three are discplayed on start-up, unless \code{all} is set to
\code{TRUE}. Default is \code{FALSE}.}

\item{...}{Additional parameters that can be used to choose the
dimentionality reduction method, as defined in
\code{\link{plot2D}}.}
}
\description{
pRoloc interactive visualisation
}
\examples{
library("pRoloc")
library("pRolocdata")
data(dunkley2006)
## markers matrix ecoding
dunkley2006 <- mrkVecToMat(dunkley2006)
## order the fractions
dunkley2006 <- dunkley2006[, order(dunkley2006$fraction)]
if (interactive())
  pRolocGUI:::pRolocVis2(dunkley2006)
}
\author{
Laurent Gatto
}

