\name{pRolocGUI-package}
\alias{pRolocGUI-package}
\docType{package}
\title{
An interactive visualisation tool for organelle (spatial) proteomics data
}
\description{
The package pRolocGUI comprises functions to interactively visualise
organelle (spatial) proteomics data on the basis of pRoloc, pRolocdata
and shiny. 
}
\details{
\tabular{ll}{
Package: \tab pRolocGUI\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-04-28\cr
License: \tab GPL-2\cr
}
}
\author{
Thomas Naake and Laurent Gatto

Maintainer: Thomas Naake <thomasnaake@gmx.de>
}
\keyword{ package }
\examples{
\dontrun{
pRolocGUI()
}
}
