\name{pRolocGUI}
\alias{pRolocGUI}
\title{
pRolocGUI
}
\description{
A function to start a shiny session with one MSnSet data set. Run
\code{pRolocGUI()} to start the shiny application and choose between
three example MSnSet originating from \code{pRolocdata} or upload your
own MSnSet. Choosing between the tabs allows to display PCA plots,
protein profiles, the underlying data and upload abilities for past
search results.

}
\usage{
pRolocGUI()
}
\author{
Thomas Naake <thomasnaake@gmx.de>
}
\examples{
\dontrun{pRolocGUI()}
}
