% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRolocVis.R, R/pRolocVis_aggregation.R,
%   R/pRolocVis_classify.R, R/pRolocVis_compare.R, R/pRolocVis_pca.R
\name{pRolocVis}
\alias{pRolocVis}
\alias{pRolocVis_aggregate}
\alias{pRolocVis_classify}
\alias{pRolocVis_compare}
\alias{pRolocVis_pca}
\title{Interactive visualisation of spatial proteomics data}
\usage{
pRolocVis(object, app = "pca", fcol, ...)

pRolocVis_aggregate(object, fcol, groupBy, fig.height = "600px",
  fig.width = "100\%", legend.width = "200\%", legend.cex = 1,
  nchar = 40, all = TRUE, mirrorX = FALSE, mirrorY = FALSE, ...)

pRolocVis_classify(object, fcol, scol, mcol = "markers",
  legend.cex = 1, ...)

pRolocVis_compare(object, fcol1, fcol2, foi, fig.height = "600px",
  fig.width = "100\%", legend.width = "200\%", legend.cex = 1,
  remap = TRUE, nchar = 40, all = TRUE, mirrorX = FALSE,
  mirrorY = FALSE, ...)

pRolocVis_pca(object, fcol = "markers", foi, fig.height = "600px",
  fig.width = "100\%", nchar = 40, all = TRUE, ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}, or an
\code{MSnSetList} of length 2 if using \code{"compare"}
application.}

\item{app}{The type of application requested: \code{"pca"}
(default), \code{"classify"}, \code{"compare"} or
\code{"aggregate"}. See description below.}

\item{fcol}{The feature meta-data label (\code{fData} column name)
to be used for colouring. Default is \code{"markers"}. This
will correspond to the prediction column if using "classify",
or the markers (labelled data) to be plotted otherwise. If set
to \code{NULL}, no annotation is expected.}

\item{...}{Additional parameters passed to \code{plot2D} for the
\code{"pca"} (such as a different dimensionality reduction
technique than PCA, and its arguments), \code{"classify"},
\code{"compare"} apps. For the \code{"aggregate"} app this is
for additional parameters to be passed to
\code{combineFeatures}.}

\item{groupBy}{The feature meta-data label (\code{fData} column name)
to be used for summarising the features to be combined.}

\item{fig.height}{Height of the figure. Default is \code{"600px"}.}

\item{fig.width}{Width of the figure. Default is \code{"100px"}.}

\item{legend.width}{Width of the legend. Default is
\code{"200\%"}.}

\item{legend.cex}{Point character expansion for the the legend.
Default is 1.}

\item{nchar}{Maximum number of characters of the markers class
names, before their names are truncated. Default is 10.}

\item{all}{If \code{TRUE} all clusters are displayed on startup,
if the total number of clusters is less than including 15. If
\code{FALSE} or otherwise, only the first cluster in the list
is displayed.}

\item{mirrorX}{Should the first PC of the second \code{MSnSet} in
\code{object} be mirrored (default is \code{FALSE}). Only
relevant when \code{remap} is \code{FALSE}.}

\item{mirrorY}{Should the second PC of the second \code{MSnSet} in
\code{object} be mirrored (default is \code{FALSE}). Only
relevant when \code{remap} is \code{FALSE}.}

\item{scol}{The feature meta data column containing the 
classification scores.}

\item{mcol}{The feature meta data column containing the labelled training 
data, for use with "classify".}

\item{fcol1}{In yhe \code{compare} app this is the feature
meta-data label (fData column name) for the first dataset in
the \code{MSnSetList}.  Default is \code{markers}.}

\item{fcol2}{In the \code{compare} app this is the feature
meta-data label (fData column name) for the second dataset in
the \code{MSnSetList}.  Default is \code{markers}.}

\item{foi}{A \code{\link{FeaturesOfInterest}} or
\code{\link{FoICollection}} object.}

\item{remap}{A \code{logical} indicating whether the second
dataset in the \code{MSnSetList} should be remapped to the
first dataset. The default is TRUE.}
}
\value{
For the \code{classify} app a \code{numeric} vector of thresholds, 
one per class, to use with \code{\link{getPredictions}}

For the \code{pca}, \code{compare} and \code{aggregate} apps 
    a \code{character} vector of \code{featureNames} names of the
    \code{object} loaded that have been selected in the app
    upon application closure.
}
\description{
These functions allow one to explore spatial proteomics data
interactively.
}
\details{
The function \code{pRolocVis} is a wrapper for
\code{pRolocVis_pca}, \code{pRolocVis_classify},\code{pRolocVis_compare}.
and \code{pRolocVis_aggregate}. These Shiny apps allow to explore and
analyse interactively spatial proteomics data.

The \code{pca} Shiny app allows exploration of quantitative data
(1) visually through Principle Component Analysis (PCA), (2)
protein profiles, and (3) a searchable feature data table,
allowing visualisation of particular proteins of interest. (NOTE:
This
\href{https://github.com/ComputationalProteomicsUnit/pRolocGUI/issues/92}{issue
on GitHub} shows how to use pre-computed coodinates prior to
visualisation with the app.)

The \code{classify} Shiny app is used to visualise classification
results and set user-specified thresholds for sub-cellular
location predictions.

The \code{compare} Shiny app is meant for comparing protein
localisation between two conditions, or two different experiments,
replicates etc. Please note that passing the argument \code{method}
to \code{...} will not work as it is already specified internally.

The \code{aggregation} Shiny app displays a scatter plot of the
maximum or mean distances within each feature (e.g. protein group)
according to its components (e.g. peptides) defined by the
\code{groupBy} argument. A PCA plot of the components is also
displayed. It can be used for visualising peptides, PSMs or any
other features defined in the feature data of the \code{MSnSet}
and their distributions.
}
\examples{
library("pRoloc")
library("pRolocdata")
data(hyperLOPIT2015)
## Load the PCA app
if (interactive()) {
  pRolocVis(hyperLOPIT2015)
}

## Load classification results from hyperLOPIT stored in fData
if (interactive()) {
  myThreshold <- pRolocVis(hyperLOPIT2015, app = "classify",
                           fcol = "svm.classification",
                           scol = "svm.score")
  newPredictions <- getPredictions(hyperLOPIT2015, fcol = "svm.classification",
                                   scol = "svm.score", t = myThreshold)
}

## Visualise the location and distribution of peptides per protein group
data("hyperLOPIT2015ms2psm")
if (interactive()) {
  ## Combine PSM data to peptides
  hl <- combineFeatures(hyperLOPIT2015ms2psm,
                        groupBy = fData(hyperLOPIT2015ms2psm)$Sequence,
                        fun = median)
  ## Visualise peptides according to protein group
  pRolocVis(hl, app = "aggregate", fcol = "markers",
            groupBy = "Protein.Group.Accessions")
}
}
\seealso{
The package vignette: \code{vignette("pRolocGUI")}.
}
\author{
Laurent Gatto, Lisa Breckels and Thomas Naake
}
