% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pRolocVis.R, R/pRolocVis_classify.R, R/pRolocVis_compare.R, R/pRolocVis_legacy.R, R/pRolocVis_pca.R, R/pRolocVis_profiles.R
\name{pRolocVis}
\alias{pRolocVis}
\alias{pRolocVis_classify}
\alias{pRolocVis_compare}
\alias{pRolocVis_legacy}
\alias{pRolocVis_pca}
\alias{pRolocVis_profiles}
\title{Interactive visualisation of `pRoloc` data}
\usage{
pRolocVis(object, what, fcol, legend.cex = 1, ...)

pRolocVis_classify(object, fcol, scol, mcol = "markers", legend.cex = 1,
  method, ...)

pRolocVis_compare(object, method = "PCA")

pRolocVis_legacy(object, method = "PCA")

pRolocVis_pca(object, fcol, foi, fig.height = "600px", fig.width = "100\%",
  legend.width = "100\%", legend.cex = 1, nchar = 15, all = TRUE,
  method)

pRolocVis_profiles(object, fcol, legend.cex = 1, all = TRUE, method, ...)
}
\arguments{
\item{object}{An instance of class \code{MSnSet}, or a list of \code{MSnSet}
objects of length 2 if using "compare" application.}

\item{what}{The type of application requested: "pca", "profiles",
"classify", "compare", "legacy". The default is "pca". See description below.}

\item{fcol}{The feature meta-data label (fData column name). This will correspond
to the prediction column if using "classify", or the markers (labelled data)
to be plotted otherwise.}

\item{legend.cex}{Character expansion for the vignette
labels. Default is 1.}

\item{...}{Additional parameters.}

\item{scol}{The feature meta data column containing the
classification scores.}

\item{mcol}{The feature meta data column containing the labelled training
data, for use with "classify".}

\item{method}{An instance of class \code{matrix} where its
rownames must match the object's feature names and represent
a projection of the data in object in two dimensions, as
produced (and invisibly returned) by \code{plot2D}.}

\item{foi}{A \code{\link{FeaturesOfInterest}} or a}

\item{fig.height}{Height of the figure. Default is \code{"600px"}.}

\item{fig.width}{Width of the figure. Default is \code{"600px"}.}

\item{legend.width}{Width of the legend. Default is \code{"100\%"}.}

\item{nchar}{Maximum number of characters of the markers class
names, before their names are truncated. Default is 10.}

\item{all}{If there are more than 10 clusters, only the first
three are discplayed on start-up, unless \code{all} is set to
\code{TRUE}. Default is \code{FALSE}.
\code{\link{FoICollection}}, that will be available for display.}
}
\value{
For \code{classify} a vector of thresholds, one per class, to use with
\code{\link{getPredictions}}

For \code{legacy} an object \code{pRolocGUI_SearchResults} of class
\code{FoICollection} when the object existed already or when a new
\code{FoICollection} was created during a session.
}
\description{
These functions allow one to explore spatial proteomics data interactively.


}
\details{
The function \code{pRolocVis} is a wrapper for \code{pRolocVis_pca},
\code{pRolocVis_profiles}, \code{pRolocVis_classification},
\code{pRolocVis_compare} and \code{pRolocVis_legacy}.These Shiny apps
allow to explore and analyse interactively spatial proteomics data.

The \code{pca} Shiny app allows exploration of quantitatuve data (1) visually
through Principle Component Analysis (PCA), (2) protein profiles, and (3)
a searchable feature data table, allowing visualisation of particualr proteins
of interest.

The \code{profiles} Shiny app allows one to simultaneously view protein
profiles and PCA locations of sub-cellular annotated sets of proteins.

The \code{classify} Shiny app is used to visualise classification results
and set user-specified thresholds for sub-cellular location predictions.

The \code{legacy} Shiny app is the original (legacy) \code{pRolocVis} app
that appears in older versions of \code{pRolocGUI}, the \code{pca} app has
been designed to replace this app.

The \code{compare} Shiny app is meant for comparing protein localisation
between two conditions, or two different experiments, replicates etc.
}
\examples{
library("pRoloc")
library("pRolocdata")
data(hyperLOPIT2015)
if (interactive())
  pRolocVis(hyperLOPIT2015, what = "pca")
  pRolocVis(hyperLOPIT2015, what = "profiles")
## Classification
opt <- knnOptimisation(hyperLOPIT2015, times = 10)
res <- knnClassification(hyperLOPIT2015, opt)
if (interactive())
  myThreshold <- pRolocVis(res, what = "classify", fcol = "knn")
  newPredictions <- getPredictions(res, fcol = "knn", t = myThreshold)
}
\author{
Laurent Gatto, Lisa Breckels and Thomas Naake
}

