\name{pRolocVIS}
\alias{pRolocVIS}
\title{pRolocVIS}
\usage{
pRolocVIS()

pRolocVIS()
}
\arguments{
  \item{object}{Pass a MSnSet to pRolocVIS directly.
  Default \code{NULL} will load no MSnSet}

  \item{object}{Pass a MSnSet to pRolocVIS directly.
  Default \code{NULL} will enable possibility to upload
  MSnSet in pRolovVIS.}
}
\description{
A function to start a shiny session with one MSnSet data
set. Run \code{pRolocVIS()} to start the shiny application
and choose between three example MSnSet originating from
\code{pRolocdata} or upload your own MSnSet. Choosing
between the tabs allows to display PCA plots, protein
profiles, the underlying data and upload abilities for past
search results.

A function to start a shiny session with one MSnSet data
set. Run \code{pRolocVIS()} to start the shiny application
and choose between three example MSnSet originating from
\code{pRolocdata} or upload your own MSnSet. Choosing
between the tabs allows to display PCA plots, protein
profiles, the underlying data and upload abilities for past
search results.
}
\examples{
\dontrun{pRolocGUI()}
\dontrun{pRolocGUI()}
}
\author{
Thomas Naake <thomasnaake@gmx.de>

Thomas Naake <thomasnaake@gmx.de>
}

