% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pRolocVis}
\alias{pRolocVis}
\title{pRolocVis}
\usage{
pRolocVis(object)
}
\arguments{
\item{object}{an object of class \code{MSnSet} or a list of \code{MSnSet}s.}
}
\value{
An object \code{pRolocGUI_SearchResults} of class \code{FoICollection}
when the object existed already or when a new \code{FoICollection} was
created during a session.
}
\description{
A function to start a shiny session with one MSnSet data set or
a list of \code{MSnSet}s to explore and analyse interactively spatial
proteomics data. \code{pRolocVis} offers high interactivity for exploring
Principle Component Analysis plots, protein profile plots and quantatative
and qualitative meta-data. Additionally, \code{pRolocVis} supports
import/export abilities for past and new search results using the
\code{FeaturesOfInterest}/\code{FoICollection} infrastructure defined in the
\code{MSnbase} package.
}
\examples{
\dontrun{

## load MSnSet data sets from the pRolocdata package
data(andy2011, package = "pRolocdata")
data(tan2009r1, package = "pRolocdata")
data(dunkley2006, package = "pRolocdata")

## create lists with unnamed and named objects
unnamed <- list(andy2011, tan2009r1, dunkley2006)
named <- list(andy2011 = andy2011, tan2009r1 = tan2009r1, dunkley2006 = dunkley2006)

## launch application by either assigning a MSnSet, an unnamed or a
## named list to the argument object
pRolocVis(object = andy2011)
pRolocVis(object = unnamed)
pRolocVis(object = named)
}
}
\author{
Thomas Naake <tn299@cam.ac.uk>
}

